pyfucktendo/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyfucktendo"
version = "1337"
dependencies = [
    "pycryptodome==3.23.0"
]

[tool.hatch.build.targets.wheel]
packages = ["pyfucktendo"]
